wip: created a utils file to hold helper functions
