+++ gkr/src/lib.rs
@@ -1,2 +1,3 @@
 pub mod gkr_protocol;
 pub mod succinct_gkr_protocol;
+pub mod utils;
